backend/pyproject.toml
[project]
name = "MyFileServer"
version = "1.1.2"
requires-python = "~=3.12.10"
dependencies = [
    "flask~=3.1.1",
    "moviepy~=2.2.1",
    "mutagen~=1.47.0",
    "pillow~=11.3.0",
    "pyperclip~=1.11.0",
    "requests~=2.32.4",
    "waitress~=3.0.2",
]

[dependency-groups]
dev = [
    "flask-cors~=6.0.1",
]
